requirements.txt
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.11.2
drf-yasg==1.20.0
geographiclib==1.52
geopy==2.2.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
parsedatetime==2.6
pexpect==4.8.0
Pillow==9.1.1
